requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
requests==2.31.0
flake8==6.0.0
black==23.3.0